+++ nodemon.json
@@ -1,0 +1,3 @@
+{
+    "ignore": ["data"]
+}
