+++ package.json
@@ -21,7 +21,8 @@
   },
   "scripts": {
     "lint": "eslint .",
-    "start": "nodemon index.js",
+    "start": "nodemon src/index.js",
+    "update-pantry": "nodemon src/pasta.js",
     "production": "pm2 start index.js",
     "stop": "pm2 stop index.js",
     "restart": "pm2 restart index.js"
